housing-consumer/create_database/requirements.txt
Flask
psycopg2-binary
Flask-Migrate
SQLAlchemy